pyproject.toml
[project]
name = "dbrik-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langchain-mcp-adapters",
    "langchain-openai",
    "langgraph",
    "python-dotenv",
    "setuptools"
]

